+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Fixed
 - Self updater crash when not launching in same working directory (Issue #224)
+- NanaZip access denied error (Issue #228)
 
 ## [1.19.0] - 2024-06-13
 ### Added