+++ CHANGELOG.md
@@ -11,7 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Self updater crash when not launching in same working directory (Issue #224)
 - NanaZip access denied error (Issue #228)
-- Minimal install asking for files to override
+- Always override existing files when performing minimal install
+
+### Changed
+- Migrated to x64 built code, which means you will have to install the runtime again. This time a dialog will appear, and will assist you
 
 ## [1.19.0] - 2024-06-13
 ### Added